requirements.txt
geopy==2.2.0
requests==2.27.1
numpy==1.22.1
tensorflow==2.8.0rc1
tensorflow-hub==0.12.0
tensorflow-object-detection-api==0.1.1
Pillow==9.0.0